feat: comprehensive /src directory cleanup and reorganization
## Major Cleanup (45% reduction in /src files):

### 🗑️ Deleted Unused Files:
- **Assets (16 removed)**: badge images, duplicate profile pics, metadata files
- **Data (3 removed + folder)**: projects.ts, experience.ts, creativeWork.ts
- **Types (1 removed + folder)**: index.ts
- **Utils (3 removed)**: microInteractions.ts, performance.ts, useResponsive.ts
- **Hooks (1 removed + folder)**: useSiteContent.ts
- **Context (1 removed)**: GitHubAuthContext.tsx
- **Content (1 removed + folder)**: site.json
- **Docs (1 removed)**: 3d/README.md

### 📁 Folder Consolidation:
- Merged /contexts into /context (moved VisualizationContext.tsx)
- Updated import path in App.tsx

### ✅ Final Clean Structure:

**Result**: 24 essential files, highly maintainable structure
