+++ src/App.tsx
@@ -1,5 +1,5 @@
 import SlideshowPortfolio from '@components/SlideshowPortfolio'
-import { VisualizationProvider } from './contexts/VisualizationContext'
+import { VisualizationProvider } from './context/VisualizationContext'
 import { CardVariantProvider } from './context/CardVariantContext'
 
 export default function App() {
