--- src/content/site.json
@@ -1,226 +1,0 @@
-{
-  "meta": {
-    "title": "Hartley H. Leroy - Creative Technologist & AI Engineer",
-    "description": "I build intelligent systems that merge AI, robotics, and data engineering. Real-time AI for autonomous robotics and cloud infrastructure.",
-    "keywords": [
-      "AI Engineer",
-      "Robotics",
-      "Cloud Infrastructure",
-      "Creative Technology",
-      "ROS2",
-      "AWS",
-      "Machine Learning"
-    ],
-    "author": "Hartley H. Leroy",
-    "resumeFile": "Hartley_LeRoy_Resume_Aug25.docx.pdf"
-  },
-  "navigation": {
-    "logo": "H.H.L",
-    "links": [
-      { "label": "About", "href": "#about" },
-      { "label": "Process", "href": "#process" },
-      { "label": "Contact", "href": "#contact" }
-    ],
-    "cta": {
-      "label": "Resume",
-      "href": "/docs/Hartley_LeRoy_Resume_Aug25.docx.pdf",
-      "external": true
-    }
-  },
-  "hero": {
-    "headline": "I build intelligent systems.",
-    "subhead": "I design and deploy real-time AI for robotics and cloud infrastructure, turning sensor data into decisions with sub-100ms latency.",
-    "primaryCta": "Learn More",
-    "secondaryCta": "Hire Me",
-    "metrics": [
-      {
-        "key": "Cloud Infra",
-        "title": "$500K optimized",
-        "blurb": "Scaled AWS ingestion to 10M+ req/day at 99.9% uptime.",
-        "skill": "infra"
-      },
-      {
-        "key": "Robotics",
-        "title": "ROS2 Autonomy",
-        "blurb": "Real-time nav & CV for drone fleets.",
-        "skill": "ros2"
-      },
-      {
-        "key": "AI/ML",
-        "title": "Production pipelines",
-        "blurb": "Streaming models across 1TB/day of data.",
-        "skill": "ml"
-      }
-    ]
-  },
-  "highlights": [
-    {
-      "title": "$500K Cloud Infra Optimized",
-      "how": "Scaled AWS data ingestion with Kinesis/IoT Core.",
-      "proof": ["-32% cost", "-48% latency", "99.9% uptime"],
-      "skill": "infra"
-    },
-    {
-      "title": "Autonomous Robotics — ROS2",
-      "how": "Real-time navigation and CV for drone fleets.",
-      "proof": ["Failsafe pathing", "<50ms fusion loop", "Fleet telemetry"],
-      "skill": "ros2"
-    },
-    {
-      "title": "Production ML Pipelines",
-      "how": "Streaming models across 1TB/day sensor data.",
-      "proof": [
-        "<100ms inference",
-        "Blue/green deploys",
-        "Monitoring & alerts"
-      ],
-      "skill": "ml"
-    }
-  ],
-  "about": {
-    "positioning": "I architect systems where robotics, AI, and cloud meet.",
-    "bestAt": [
-      "Real-time data systems",
-      "Autonomous navigation",
-      "Production ML pipelines"
-    ],
-    "avoid": ["Brand design", "Adtech", "Web3"],
-    "bio": [
-      "I'm a creative technologist passionate about building intelligent systems that merge AI, robotics, and data engineering. Currently at Lucid Bots, I architect autonomous drone fleets with real-time machine learning capabilities that revolutionize how robots perceive and interact with the world.",
-      "My unique approach combines creative problem-solving with cutting-edge technology. From engineering $500K AWS IoT infrastructures to designing ROS2-based autonomous systems, I bridge the gap between visionary concepts and production-ready solutions.",
-      "I thrive at the intersection of AI innovation and practical engineering. Whether it's optimizing real-time telemetry pipelines, implementing computer vision algorithms, or leading cross-functional teams, I bring both technical expertise and creative vision to every challenge."
-    ],
-    "quote": "The future belongs to those who can bridge the gap between imagination and implementation."
-  },
-  "skills": {
-    "Ship": [
-      {
-        "name": "ROS2",
-        "evidence": "Used in 3 fleets"
-      },
-      {
-        "name": "PX4",
-        "evidence": "100+ autonomous flights"
-      },
-      {
-        "name": "Python",
-        "evidence": "Production ML pipelines"
-      },
-      {
-        "name": "C++",
-        "evidence": "Real-time robotics"
-      }
-    ],
-    "Scale": [
-      {
-        "name": "AWS IoT",
-        "evidence": "10M+ msgs/day"
-      },
-      {
-        "name": "Kinesis",
-        "evidence": "1TB+ data/day"
-      },
-      {
-        "name": "Glue",
-        "evidence": "ETL automation"
-      },
-      {
-        "name": "S3",
-        "evidence": "Petabyte storage"
-      },
-      {
-        "name": "Step Functions",
-        "evidence": "Orchestration"
-      }
-    ],
-    "Serve": [
-      {
-        "name": "Docker",
-        "evidence": "Containerized ML"
-      },
-      {
-        "name": "ECS/EKS",
-        "evidence": "Scalable deployment"
-      },
-      {
-        "name": "CI/CD",
-        "evidence": "Automated pipelines"
-      }
-    ]
-  },
-  "process": [
-    {
-      "step": "Scope",
-      "outcome": "Clarify goals and constraints.",
-      "derisk": "Prevents scope creep and misaligned timelines."
-    },
-    {
-      "step": "Prototype",
-      "outcome": "Prove architecture & latency in 2–4 weeks.",
-      "derisk": "De-risks scale and integration early."
-    },
-    {
-      "step": "Ship & Monitor",
-      "outcome": "Deploy with observability.",
-      "derisk": "Keeps reliability high post-launch."
-    }
-  ],
-  "testimonials": [
-    {
-      "quote": "Cut our streaming costs by 30% while improving latency.",
-      "author": "CTO, Robotics Co.",
-      "link": ""
-    },
-    {
-      "quote": "Hartley's ROS2 expertise transformed our autonomous systems.",
-      "author": "Lead Engineer, Drone Startup",
-      "link": ""
-    },
-    {
-      "quote": "Delivered production ML pipeline ahead of schedule.",
-      "author": "Product Manager, AI Platform",
-      "link": ""
-    }
-  ],
-  "contact": {
-    "hook": "Have a problem that needs real-time decisions?",
-    "sub": "Tell me the goal, the data you have, and the deadline. I'll outline what's realistic and how we'd ship it.",
-    "form": {
-      "name": "Your Name",
-      "email": "Your Email",
-      "subject": "Subject",
-      "message": "Your Message",
-      "submit": "Send Message",
-      "sending": "Sending...",
-      "success": "Message sent successfully! I'll get back to you soon."
-    },
-    "social": [
-      {
-        "name": "GitHub",
-        "url": "https://github.com/hartleyhleroy",
-        "icon": "github"
-      },
-      {
-        "name": "LinkedIn",
-        "url": "https://linkedin.com/in/hartleyhleroy",
-        "icon": "linkedin"
-      },
-      {
-        "name": "Email",
-        "url": "mailto:contact@hartleyleroy.dev",
-        "icon": "mail"
-      }
-    ],
-    "info": {
-      "email": "contact@hartleyleroy.dev",
-      "location": "Available Worldwide (Remote)"
-    }
-  },
-  "footer": {
-    "copyright": "© 2024 Hartley H. Leroy. All rights reserved.",
-    "links": [
-      { "label": "Privacy", "href": "/privacy" },
-      { "label": "Terms", "href": "/terms" }
-    ]
-  }
-}
